refactor: 기수 값 타입 Integer로 통일 (#324)

+++ src/main/java/org/sopt/app/application/auth/PlaygroundAuthInfo.java
@@ -50,7 +50,7 @@ public class PlaygroundAuthInfo {
     @Builder
     public static class UserActiveInfo {
 
-        private Long currentGeneration;
+        private Integer currentGeneration;
         private UserStatus status;
     }
 
@@ -87,7 +87,7 @@ public class PlaygroundAuthInfo {
         private UserStatus status;
         private String name;
         private String profileImage;
-        private List<Long> generationList;
+        private List<Integer> generationList;
     }
 
     @Getter
